Move doughnut to circular package
