+++ lib/src/commonMain/kotlin/pl/krystiankaniowski/composecharts/circural/DoughnutChart.kt
@@ -1,4 +1,4 @@
-package pl.krystiankaniowski.composecharts.pie
+package pl.krystiankaniowski.composecharts.circural
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.border
